# documentation/docs/content/openAPI/spring_backend-openapi.yaml
openapi: "3.0.3"
info:
  title: "spring_backend API"
  description: "spring_backend API"
  version: "1.0.0"
servers:
  - url: "https://spring_backend"
paths:
  /api/projects/{id}:
    put:
      summary: "PUT api/projects/{id}"
      operationId: "updateProject"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProjectDTO"
    delete:
      summary: "DELETE api/projects/{id}"
      operationId: "deleteProject"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProjectDTO"
    get:
      summary: "GET api/projects/{id}"
      operationId: "getProjectById"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProjectDTO"
  /api/projects:
    post:
      summary: "POST api/projects"
      operationId: "createNewProject"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProjectDTO"
    get:
      summary: "GET api/projects"
      operationId: "getAllProjects"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/ProjectDTO"
...
  /api/projects/{projectId}/tracks/{trackId}/datasets/{datasetId}:
    put:
      summary: "PUT api/projects/{projectId}/tracks/{trackId}/datasets/{datasetId}"
      operationId: "updateDataSet"
      parameters:
        - name: "projectId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "trackId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "datasetId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlainDataDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlainDataDTO"
    delete:
      summary: "DELETE api/projects/{projectId}/tracks/{trackId}/datasets/{datasetId}"
      operationId: "deleteDataSet"
      parameters:
        - name: "projectId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "trackId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "datasetId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlainDataDTO"
    get:
      summary: "GET api/projects/{projectId}/tracks/{trackId}/datasets/{datasetId}"
      operationId: "getDatasetById"
      parameters:
        - name: "projectId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "trackId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "datasetId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlainDataDTO"
  /api/projects/{projectId}/tracks/{trackId}/datasets:
    post:
      summary: "POST api/projects/{projectId}/tracks/{trackId}/datasets"
      operationId: "createNewDataSet"
      parameters:
        - name: "projectId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "trackId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlainDataDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlainDataDTO"
    get:
      summary: "Get all foos"
      description: ""
      operationId: ""
      parameters:
        - name: "projectId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "trackId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/PlainDataDTO"
  /api/projects/{projectId}/tracks/{trackId}/datasets/{datasetId}/preview:
    get:
      summary: "GET api/projects/{projectId}/tracks/{trackId}/datasets/{datasetId}/preview"
      operationId: "getPreview"
      parameters:
        - name: "projectId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "trackId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "datasetId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsonNode"
  /api/projects/ml/datasets/{id}:
    get:
      summary: "GET api/projects/ml/datasets/{id}"
      operationId: "getAllMLDataSets"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/PlainDataOccurrenceDTO"
  /api/projects/tsd/datasets/{id}:
    get:
      summary: "GET api/projects/tsd/datasets/{id}"
      operationId: "getAllTSDDataSets"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/PlainDataOccurrenceDTO"
...
components:
  schemas:
    ProjectDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        sending:
          type: "boolean"
...
    JsonNode:
      type: "object"
      properties: { }
...
    PlainDataDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        numSamples:
          type: "integer"
          format: "int32"
        position:
          type: "integer"
          format: "int32"
        frequency:
          type: "number"
          format: "float"
        dataType:
          type: "string"
          enum:
            - "CHAR"
            - "FLOAT"
            - "ML"
            - "TSA"
            - "INTEGER"
            - "SLEEP"
        customValues:
          $ref: "#/components/schemas/JsonNode"
    PlainDataOccurrenceDTO:
      type: "object"
      properties:
        projectId:
          type: "integer"
          format: "int64"
        trackId:
          type: "integer"
          format: "int64"
        dataSetId:
          type: "integer"
          format: "int64"
        projectName:
          type: "string"
        trackName:
          type: "string"
        dataSetName:
          type: "string"
...